requirements.txt
Django==4.1.6
requests==2.28.1
requests-oauthlib==1.3.1
django-crontab
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
environ==1.0
google-api-python-client==2.75.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
apiclient==1.0.4
isodate==0.6.1